feat: extract single children up in video device select UI
